scsi: sg: Change rwlock to spinlock
A reviewer suggested that the extra overhead associated with a rw lock
compared to a spinlock was not worth it for short, oft-used critcal
sections.

So the rwlock on the request list/array is changed to a spinlock.  The head
of that list is in the owning sf file descriptor object.

Link: https://lore.kernel.org/r/20210113224526.861000-12-dgilbert@interlog.com
Reviewed-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: Douglas Gilbert <dgilbert@interlog.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>