scsi: sg: printk() change %p to %pK
This driver does a lot of buffer juggling in an attempt to take some of
that chore away from its users. When debugging problems associated with
that buffer juggling getting sensible pointer values is a major aid. So
change %p to %pK. The system administrator can choose to obfuscate %pK
pointers. The "pK" is also easier to search for in the code if further
changes are required.

Link: https://lore.kernel.org/r/20210113224526.861000-23-dgilbert@interlog.com
Reviewed-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: Douglas Gilbert <dgilbert@interlog.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>