scsi: sg: Sense buffer rework
The biggest single item in the sg_request object is the sense buffer array
which is SCSI_SENSE_BUFFERSIZE bytes long. That constant started out at 18
bytes 20 years ago and is 96 bytes now and might grow in the future. On the
other hand the sense buffer is only used by a small number of SCSI
commands: those that fail and those that want to return more information
other than a SCSI status of GOOD.

Set up a small mempool called "sg_sense" that is only used as required and
released back to the mempool as soon as practical.

Link: https://lore.kernel.org/r/20210113224526.861000-27-dgilbert@interlog.com
Reviewed-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: Douglas Gilbert <dgilbert@interlog.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>